// src/lib/types/personalMeeting.type.ts
export type EyelidType = 'DOUBLE' | 'INNER' | 'SINGLE';

export type AppearanceType = 'ARAB' | 'NORMAL' | 'TOFU';

export type SmokingType = 'CIGARETTE' | 'E_CIGARETTE' | 'FALSE';

export type WeightType = 'HEIGHT' | 'AGE' | 'APPEARANCE' | 'SMOKING' | 'MBTI';

export type MoodType = 'ACTICE' | 'CALM';

export type DepartmentType =
  | '전자전기컴퓨터공학부'
  | '화학공학과'
  | '기계정보공학과'
  | '신소재공학과'
  | '토목공학과'
  | '컴퓨터과학부'
  | '인공지능학과'
  | '행정학과'
  | '국제관계학과'
  | '경제학부'
  | '사회복지학과'
  | '세무학과'
  | '영어영문학과'
  | '국어국문학과'
  | '국사학과'
  | '철학과'
  | '중국어문화학과'
  | '수학과'
  | '통계학과'
  | '물리학과'
  | '생명과학과'
  | '환경원예학과'
  | '융합응용화학과'
  | '도시행정학과'
  | '도시사회학과'
  | '건축학부'
  | '도시공학과'
  | '교통공학과'
  | '조경학과'
  | '환경공학부'
  | '공간정보공학과'
  | '디자인학과'
  | '조각학과'
  | '음악학과'
  | '스포츠과학과'
  | '첨단융합학부'
  | '자유전공학부'
  | '융합전공학부';